% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarshallOlkin.R
\name{BiCopEst.MO}
\alias{BiCopEst.MO}
\title{Estimation of Marshall-Olkin copulas}
\usage{
BiCopEst.MO(
  u1,
  u2,
  method,
  par = 0.5,
  kernel = "exp-l2",
  gamma = 0.3,
  alpha = 1,
  niter = 100,
  ndrawings = 10,
  naveraging = 1
)
}
\arguments{
\item{method}{a character giving the name of the estimation method, among:
\itemize{
  \item \code{curve}: \eqn{\alpha} is estimated by inversion of
    the probability measure of the diagonal
    \eqn{\{(u,v): u = v\}}{ {(u,v): u = v} }
  \item \code{itau}: \eqn{\alpha} is estimated by inversion of Kendall's tau
  \item \code{MMD}: \eqn{\alpha} is estimated by MMD optimization
}}
}
\description{
Estimation of Marshall-Olkin copulas
}
