% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarshallOlkin.R
\name{BiCopSim.MO}
\alias{BiCopSim.MO}
\title{Simulation of Marshall-Olkin copula}
\usage{
BiCopSim.MO(n, alpha)
}
\arguments{
\item{n}{number of samples}

\item{alpha}{parameter of the Marshall-Olkin copula}
}
\value{
an \eqn{n \times 2} matrix containing the samples
}
\description{
This functions simulates independent realizations
from the Marshall-Olkin copula.
}
\examples{
# Simulation from a Marshall-Olkin copula with parameter alpha = 0.5
BiCopSim.MO(n = 100, alpha = 0.5)


}
\seealso{
\code{\link{BiCopEst.MO}} for the estimation of
Marshall-Olkin copulas.
}
